+++ packages/installer/src/calls/packageRemove.ts
@@ -10,7 +10,6 @@ import {
   dockerComposeDown,
   listPackage,
 } from "@dappnode/dockerapi";
-import { isRunningHttps } from "@dappnode/httpsportal";
 import { httpsPortal } from "@dappnode/httpsportal";
 import * as db from "@dappnode/db";
 import { mevBoostMainnet, mevBoostPrater, stakerPkgs } from "@dappnode/types";
@@ -44,18 +43,7 @@ export async function packageRemove({
   // Remove portal https portal mappings if any.
   // MUST removed before deleting containers
   try {
-    if ((await isRunningHttps()) === true) {
-      const mappings = await httpsPortal.getMappings(dnp.containers);
-      for (const mapping of mappings) {
-        if (mapping.dnpName === dnpName)
-          await httpsPortal
-            .removeMapping(mapping)
-            // Bypass error to continue deleting mappings
-            .catch((e) =>
-              logs.error(`Error removing https mapping of ${dnp.dnpName}`, e)
-            );
-      }
-    }
+    httpsPortal.removeMappings(dnp);
   } catch (e) {
     // Bypass error to continue deleting the package
     logs.error(
